# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "API",
    "termsOfService": "http://swagger.io/terms/"
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/level3": {
      "get": {
        "tags": [
          "level3"
        ],
        "summary": "Get all the level3 objects",
        "description": "",
        "operationId": "getLevel3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "post": {
        "tags": [
          "level3"
        ],
        "summary": "Add a new level3 item",
        "description": "",
        "operationId": "addLevel3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Level3 object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Level3"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/level3/{id}": {
      "get": {
        "tags": [
          "level3"
        ],
        "summary": "Get level3 object by ID",
        "operationId": "getLevel3ById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of object to return",
            "required": true,
            "type": "string",
            "format": "ObjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Level3"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Object not found"
          }
        }
      },
      "post": {
        "tags": [
          "level3"
        ],
        "summary": "Update level3 object",
        "description": "",
        "operationId": "updateLevel3item",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Level3 object that needs to be added to the store",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "level3"
        ],
        "summary": "Deletes a level3 object",
        "description": "",
        "operationId": "deletePet",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "petId",
            "in": "path",
            "description": "Pet id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:pets",
              "read:pets"
            ]
          }
        ]
      }
    }
  }
}